requirements.txt
alembic==1.3.2
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
bandit==1.6.2
bcrypt==3.1.7
black==19.10b0
cffi==1.13.2
Click==7.0
colorama==0.4.3
colorlog==4.1.0
coverage==5.0.1
decorator==4.4.1
entrypoints==0.3
flake8==3.7.9
flake8-polyfill==1.0.2
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-Testing==0.7.1
future==0.18.2
gitdb2==2.0.6
GitPython==2.1.14
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.2.0
jupyter-core==4.6.1
lazy-object-proxy==1.4.3
Mako==1.1.0
mando==0.6.4
MarkupSafe==1.1.1
marshmallow==3.3.0
mccabe==0.6.1
nbformat==4.4.0
pathspec==0.7.0
pbr==5.4.4
plotly==4.5.0
progress==1.4
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.4.4
pyrsistent==0.15.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
radon==4.0.0
regex==2020.1.8
retrying==1.3.3
six==1.13.0
smmap2==2.0.5
SQLAlchemy==1.3.12
stevedore==1.31.0
tabulate==0.8.6
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
Werkzeug==0.16.0
wily==1.16.0
wrapt==1.11.2
